@@ -17,7 +17,7 @@
     "cy:run": "start-server-and-test 'yarn website:test' http://localhost:3000 'cypress run --headless'",
     "cy:verify": "cypress verify",
     "lint:css": "stylelint **/src/**/*.css",
-    "lint": "eslint --ext .js,.ts,.tsx .",
+    "lint": "eslint .",
     "playground:build": "yarn workspace @docsearch/react-example build",
     "playground:start": "yarn workspace @docsearch/react-example dev",
     "playground-js:start": "yarn workspace @docsearch/js-example dev",
@@ -25,7 +25,7 @@
     "start": "yarn run watch",
     "test:size": "bundlesize",
     "test:types": "tsc --noEmit",
-    "test": "jest",
+    "test": "vitest",
     "watch": "lerna run watch --parallel",
     "website:build": "yarn workspace @docsearch/website build",
     "website:start": "yarn workspace @docsearch/website start --host 0.0.0.0",
@@ -35,41 +35,38 @@
     "@babel/plugin-transform-react-jsx": "7.23.4",
     "@babel/preset-env": "7.23.5",
     "@babel/preset-typescript": "7.23.3",
+    "@eslint-community/eslint-plugin-eslint-comments": "4.4.1",
     "@rollup/plugin-babel": "^6.0.4",
     "@rollup/plugin-json": "6.1.0",
     "@rollup/plugin-node-resolve": "15.3.0",
     "@rollup/plugin-replace": "6.0.1",
     "@rollup/plugin-terser": "0.4.4",
+    "@stylistic/eslint-plugin": "2.10.1",
     "@testing-library/dom": "9.3.1",
-    "@testing-library/jest-dom": "5.16.5",
-    "@testing-library/react": "14.1.2",
-    "@types/jest": "^29.0.0",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
-    "@typescript-eslint/eslint-plugin": "5.41.0",
-    "@typescript-eslint/parser": "5.41.0",
+    "@typescript-eslint/eslint-plugin": "8.12.2",
+    "@typescript-eslint/parser": "8.12.2",
     "babel-plugin-inline-replace-variables": "1.3.1",
     "babel-plugin-module-resolver": "4.0.0",
     "bundlesize2": "0.0.31",
     "cssnano": "4.1.11",
     "cypress": "13.15.1",
-    "eslint": "8.55.0",
-    "eslint-config-algolia": "20.0.0",
-    "eslint-config-prettier": "8.10.0",
+    "eslint": "8.57.1",
+    "eslint-config-algolia": "23.2.1",
+    "eslint-config-prettier": "9.1.0",
     "eslint-plugin-cypress": "4.1.0",
-    "eslint-plugin-eslint-comments": "3.2.0",
-    "eslint-plugin-import": "2.29.0",
-    "eslint-plugin-jest": "27.6.0",
-    "eslint-plugin-jsdoc": "39.4.0",
-    "eslint-plugin-jsx-a11y": "6.6.1",
-    "eslint-plugin-prettier": "4.2.1",
-    "eslint-plugin-react": "7.31.10",
-    "eslint-plugin-react-hooks": "4.6.0",
-    "jest": "29.6.1",
-    "jest-environment-jsdom": "29.6.1",
+    "eslint-plugin-import": "2.31.0",
+    "eslint-plugin-jsdoc": "50.4.3",
+    "eslint-plugin-jsx-a11y": "6.10.2",
+    "eslint-plugin-prettier": "5.2.1",
+    "eslint-plugin-react": "7.37.2",
+    "eslint-plugin-react-hooks": "5.0.0",
+    "eslint-plugin-react-refresh": "0.4.14",
+    "jsdom": "25.0.1",
     "lerna": "8.1.9",
     "postcss": "8.4.25",
-    "prettier": "2.7.1",
+    "prettier": "3.3.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "rollup": "4.24.4",
@@ -79,13 +76,13 @@
     "start-server-and-test": "1.15.5",
     "stylelint": "14.16.1",
     "stylelint-a11y": "1.2.3",
-    "stylelint-config-sass-guidelines": "9.0.1",
-    "stylelint-config-standard": "28.0.0",
-    "stylelint-no-unsupported-browser-features": "5.0.4",
-    "stylelint-order": "5.0.0",
-    "stylelint-prettier": "3.0.0",
-    "ts-jest": "28.0.8",
+    "stylelint-config-sass-guidelines": "12.1.0",
+    "stylelint-config-standard": "36.0.1",
+    "stylelint-no-unsupported-browser-features": "8.0.2",
+    "stylelint-order": "6.0.4",
+    "stylelint-prettier": "5.0.2",
     "typescript": "5.6.3",
+    "vitest": "^2.1.4",
     "watch": "1.0.2"
   },
   "packageManager": "yarn@4.5.1"
