@@ -1,3 +1,3 @@
-module.exports = {
+export default {
   presets: [require.resolve('@docusaurus/core/lib/babel/preset')],
 };
