@@ -2,10 +2,7 @@ import type { DocSearchProps as DocSearchComponentProps } from '@docsearch/react
 import { DocSearch, version } from '@docsearch/react';
 import React, { render } from 'preact/compat';
 
-function getHTMLElement(
-  value: HTMLElement | string,
-  environment: DocSearchProps['environment'] = window
-): HTMLElement {
+function getHTMLElement(value: HTMLElement | string, environment: DocSearchProps['environment'] = window): HTMLElement {
   if (typeof value === 'string') {
     return environment.document.querySelector<HTMLElement>(value)!;
   }
@@ -18,18 +15,16 @@ interface DocSearchProps extends DocSearchComponentProps {
   environment?: typeof window;
 }
 
-export function docsearch(props: DocSearchProps) {
+export function docsearch(props: DocSearchProps): void {
   render(
     <DocSearch
       {...props}
       transformSearchClient={(searchClient) => {
         searchClient.addAlgoliaAgent('docsearch.js', version);
 
-        return props.transformSearchClient
-          ? props.transformSearchClient(searchClient)
-          : searchClient;
+        return props.transformSearchClient ? props.transformSearchClient(searchClient) : searchClient;
       }}
     />,
-    getHTMLElement(props.container, props.environment)
+    getHTMLElement(props.container, props.environment),
   );
 }
