@@ -15,65 +15,49 @@ export type DocSearchButtonProps = React.ComponentProps<'button'> & {
 const ACTION_KEY_DEFAULT = 'Ctrl' as const;
 const ACTION_KEY_APPLE = '⌘' as const;
 
-function isAppleDevice() {
+function isAppleDevice(): boolean {
   return /(Mac|iPhone|iPod|iPad)/i.test(navigator.platform);
 }
 
-export const DocSearchButton = React.forwardRef<
-  HTMLButtonElement,
-  DocSearchButtonProps
->(({ translations = {}, ...props }, ref) => {
-  const { buttonText = 'Search', buttonAriaLabel = 'Search' } = translations;
+export const DocSearchButton = React.forwardRef<HTMLButtonElement, DocSearchButtonProps>(
+  ({ translations = {}, ...props }, ref) => {
+    const { buttonText = 'Search', buttonAriaLabel = 'Search' } = translations;
 
-  const [key, setKey] = useState<
-    typeof ACTION_KEY_APPLE | typeof ACTION_KEY_DEFAULT | null
-  >(null);
+    const [key, setKey] = useState<typeof ACTION_KEY_APPLE | typeof ACTION_KEY_DEFAULT | null>(null);
 
-  useEffect(() => {
-    if (typeof navigator !== 'undefined') {
-      isAppleDevice() ? setKey(ACTION_KEY_APPLE) : setKey(ACTION_KEY_DEFAULT);
-    }
-  }, []);
+    useEffect(() => {
+      if (typeof navigator !== 'undefined') {
+        isAppleDevice() ? setKey(ACTION_KEY_APPLE) : setKey(ACTION_KEY_DEFAULT);
+      }
+    }, []);
 
-  return (
-    <button
-      type="button"
-      className="DocSearch DocSearch-Button"
-      aria-label={buttonAriaLabel}
-      {...props}
-      ref={ref}
-    >
-      <span className="DocSearch-Button-Container">
-        <SearchIcon />
-        <span className="DocSearch-Button-Placeholder">{buttonText}</span>
-      </span>
+    return (
+      <button type="button" className="DocSearch DocSearch-Button" aria-label={buttonAriaLabel} {...props} ref={ref}>
+        <span className="DocSearch-Button-Container">
+          <SearchIcon />
+          <span className="DocSearch-Button-Placeholder">{buttonText}</span>
+        </span>
 
-      <span className="DocSearch-Button-Keys">
-        {key !== null && (
-          <>
-            <DocSearchButtonKey
-              reactsToKey={
-                key === ACTION_KEY_DEFAULT ? ACTION_KEY_DEFAULT : 'Meta'
-              }
-            >
-              {key === ACTION_KEY_DEFAULT ? <ControlKeyIcon /> : key}
-            </DocSearchButtonKey>
-            <DocSearchButtonKey reactsToKey="k">K</DocSearchButtonKey>
-          </>
-        )}
-      </span>
-    </button>
-  );
-});
+        <span className="DocSearch-Button-Keys">
+          {key !== null && (
+            <>
+              <DocSearchButtonKey reactsToKey={key === ACTION_KEY_DEFAULT ? ACTION_KEY_DEFAULT : 'Meta'}>
+                {key === ACTION_KEY_DEFAULT ? <ControlKeyIcon /> : key}
+              </DocSearchButtonKey>
+              <DocSearchButtonKey reactsToKey="k">K</DocSearchButtonKey>
+            </>
+          )}
+        </span>
+      </button>
+    );
+  },
+);
 
 type DocSearchButtonKeyProps = {
   reactsToKey?: string;
 };
 
-function DocSearchButtonKey({
-  reactsToKey,
-  children,
-}: React.PropsWithChildren<DocSearchButtonKeyProps>) {
+function DocSearchButtonKey({ reactsToKey, children }: React.PropsWithChildren<DocSearchButtonKeyProps>): JSX.Element {
   const [isKeyDown, setIsKeyDown] = useState(false);
 
   useEffect(() => {
@@ -81,13 +65,13 @@ function DocSearchButtonKey({
       return undefined;
     }
 
-    function handleKeyDown(e: KeyboardEvent) {
+    function handleKeyDown(e: KeyboardEvent): void {
       if (e.key === reactsToKey) {
         setIsKeyDown(true);
       }
     }
 
-    function handleKeyUp(e: KeyboardEvent) {
+    function handleKeyUp(e: KeyboardEvent): void {
       if (
         e.key === reactsToKey ||
         // keyup doesn't fire when Command is held down,
@@ -102,20 +86,14 @@ function DocSearchButtonKey({
     window.addEventListener('keydown', handleKeyDown);
     window.addEventListener('keyup', handleKeyUp);
 
-    return () => {
+    return (): void => {
       window.removeEventListener('keydown', handleKeyDown);
       window.removeEventListener('keyup', handleKeyUp);
     };
   }, [reactsToKey]);
 
   return (
-    <kbd
-      className={
-        isKeyDown
-          ? 'DocSearch-Button-Key DocSearch-Button-Key--pressed'
-          : 'DocSearch-Button-Key'
-      }
-    >
+    <kbd className={isKeyDown ? 'DocSearch-Button-Key DocSearch-Button-Key--pressed' : 'DocSearch-Button-Key'}>
       {children}
     </kbd>
   );
