@@ -11,11 +11,9 @@ type ErrorScreenProps = {
   translations?: ErrorScreenTranslations;
 };
 
-export function ErrorScreen({ translations = {} }: ErrorScreenProps) {
-  const {
-    titleText = 'Unable to fetch results',
-    helpText = 'You might want to check your network connection.',
-  } = translations;
+export function ErrorScreen({ translations = {} }: ErrorScreenProps): JSX.Element {
+  const { titleText = 'Unable to fetch results', helpText = 'You might want to check your network connection.' } =
+    translations;
   return (
     <div className="DocSearch-ErrorScreen">
       <div className="DocSearch-Screen-Icon">
