@@ -1,7 +1,4 @@
-import type {
-  AutocompleteApi,
-  AutocompleteState,
-} from '@algolia/autocomplete-core';
+import type { AutocompleteApi, AutocompleteState } from '@algolia/autocomplete-core';
 import type { MutableRefObject } from 'react';
 import React from 'react';
 
@@ -20,12 +17,7 @@ export type SearchBoxTranslations = Partial<{
 }>;
 
 interface SearchBoxProps
-  extends AutocompleteApi<
-    InternalDocSearchHit,
-    React.FormEvent,
-    React.MouseEvent,
-    React.KeyboardEvent
-  > {
+  extends AutocompleteApi<InternalDocSearchHit, React.FormEvent, React.MouseEvent, React.KeyboardEvent> {
   state: AutocompleteState<InternalDocSearchHit>;
   autoFocus: boolean;
   inputRef: MutableRefObject<HTMLInputElement | null>;
@@ -34,7 +26,7 @@ interface SearchBoxProps
   translations?: SearchBoxTranslations;
 }
 
-export function SearchBox({ translations = {}, ...props }: SearchBoxProps) {
+export function SearchBox({ translations = {}, ...props }: SearchBoxProps): JSX.Element {
   const {
     resetButtonTitle = 'Clear the query',
     resetButtonAriaLabel = 'Clear the query',
@@ -69,9 +61,7 @@ export function SearchBox({ translations = {}, ...props }: SearchBoxProps) {
       >
         <label className="DocSearch-MagnifierLabel" {...props.getLabelProps()}>
           <SearchIcon />
-          <span className="DocSearch-VisuallyHiddenForAccessibility">
-            {searchInputLabel}
-          </span>
+          <span className="DocSearch-VisuallyHiddenForAccessibility">{searchInputLabel}</span>
         </label>
 
         <div className="DocSearch-LoadingIndicator">
@@ -99,12 +89,7 @@ export function SearchBox({ translations = {}, ...props }: SearchBoxProps) {
         </button>
       </form>
 
-      <button
-        className="DocSearch-Cancel"
-        type="reset"
-        aria-label={cancelButtonAriaLabel}
-        onClick={props.onClose}
-      >
+      <button className="DocSearch-Cancel" type="reset" aria-label={cancelButtonAriaLabel} onClick={props.onClose}>
         {cancelButtonText}
       </button>
     </>
