@@ -1,15 +1,14 @@
-import { render, act, fireEvent, screen } from '@testing-library/react';
+import { render, act, fireEvent, screen, cleanup } from '@testing-library/react';
 import React from 'react';
+import { describe, it, expect, afterEach } from 'vitest';
 
-import '@testing-library/jest-dom';
+import '@testing-library/jest-dom/vitest';
 
 import { DocSearch as DocSearchComponent } from '../DocSearch';
 import type { DocSearchProps } from '../DocSearch';
 
-function DocSearch(props: Partial<DocSearchProps>) {
-  return (
-    <DocSearchComponent appId="woo" apiKey="foo" indexName="bar" {...props} />
-  );
+function DocSearch(props: Partial<DocSearchProps>): JSX.Element {
+  return <DocSearchComponent appId="woo" apiKey="foo" indexName="bar" {...props} />;
 }
 
 // mock empty response
@@ -36,6 +35,10 @@ function noResultSearch(_queries: any, _requestOptions?: any): Promise<any> {
 describe('api', () => {
   const docSearchSelector = '.DocSearch';
 
+  afterEach(() => {
+    cleanup();
+  });
+
   it('renders with minimal parameters', () => {
     render(<DocSearch />);
 
@@ -52,15 +55,11 @@ describe('api', () => {
               buttonAriaLabel: 'Recherche',
             },
           }}
-        />
+        />,
       );
       expect(document.querySelector(docSearchSelector)).toBeInTheDocument();
-      expect(
-        document.querySelector('.DocSearch-Button-Placeholder')?.innerHTML
-      ).toBe('Recherche');
-      expect(
-        document.querySelector('.DocSearch-Button')?.getAttribute('aria-label')
-      ).toBe('Recherche');
+      expect(document.querySelector('.DocSearch-Button-Placeholder')?.innerHTML).toBe('Recherche');
+      expect(document.querySelector('.DocSearch-Button')?.getAttribute('aria-label')).toBe('Recherche');
     });
 
     it('overrides the default DocSearchModal startScreen text', async () => {
@@ -73,7 +72,7 @@ describe('api', () => {
               },
             },
           }}
-        />
+        />,
       );
 
       expect(document.querySelector(docSearchSelector)).toBeInTheDocument();
@@ -99,14 +98,13 @@ describe('api', () => {
             modal: {
               noResultsScreen: {
                 noResultsText: 'Pas de résultats pour',
-                reportMissingResultsText:
-                  'Ouvrez une issue sur docsearch-configs',
+                reportMissingResultsText: 'Ouvrez une issue sur docsearch-configs',
                 reportMissingResultsLinkText: 'Lien du repo',
               },
             },
           }}
           getMissingResultsUrl={() => 'algolia.com'}
-        />
+        />,
       );
 
       expect(document.querySelector(docSearchSelector)).toBeInTheDocument();
@@ -122,13 +120,11 @@ describe('api', () => {
       });
 
       expect(screen.getByText(/Pas de résultats pour/)).toBeInTheDocument();
-      expect(
-        screen.getByText(/Ouvrez une issue sur docsearch-configs/)
-      ).toBeInTheDocument();
+      expect(screen.getByText(/Ouvrez une issue sur docsearch-configs/)).toBeInTheDocument();
       expect(
         screen.getByRole('link', {
           name: 'Lien du repo',
-        })
+        }),
       ).toBeInTheDocument();
     });
 
@@ -146,31 +142,21 @@ describe('api', () => {
               },
             },
           }}
-        />
+        />,
       );
 
       await act(async () => {
         fireEvent.click(await screen.findByText('Search'));
       });
 
-      const searchInputLabel = document.querySelector(
-        '.DocSearch-MagnifierLabel'
-      );
+      const searchInputLabel = document.querySelector('.DocSearch-MagnifierLabel');
 
       expect(document.querySelector(docSearchSelector)).toBeInTheDocument();
 
-      expect(document.querySelector('.DocSearch-Cancel')?.innerHTML).toBe(
-        'Annuler'
-      );
-      expect(
-        document.querySelector('.DocSearch-Cancel')?.getAttribute('aria-label')
-      ).toBe('Annuler');
-      expect(
-        document.querySelector('.DocSearch-Reset')?.getAttribute('title')
-      ).toBe('Effacer');
-      expect(
-        document.querySelector('.DocSearch-Reset')?.getAttribute('aria-label')
-      ).toBe('Effacer');
+      expect(document.querySelector('.DocSearch-Cancel')?.innerHTML).toBe('Annuler');
+      expect(document.querySelector('.DocSearch-Cancel')?.getAttribute('aria-label')).toBe('Annuler');
+      expect(document.querySelector('.DocSearch-Reset')?.getAttribute('title')).toBe('Effacer');
+      expect(document.querySelector('.DocSearch-Reset')?.getAttribute('aria-label')).toBe('Effacer');
       expect(searchInputLabel?.textContent).toBe('Recherche');
     });
 
@@ -191,7 +177,7 @@ describe('api', () => {
               },
             },
           }}
-        />
+        />,
       );
 
       expect(document.querySelector(docSearchSelector)).toBeInTheDocument();
@@ -205,24 +191,14 @@ describe('api', () => {
       expect(screen.getByText('Pour naviguer')).toBeInTheDocument();
       expect(screen.getByText('Pour selectionner')).toBeInTheDocument();
       expect(
-        document.querySelector(
-          '.DocSearch-Commands-Key > svg[aria-label="Touche d\'échappement"]'
-        )
+        document.querySelector('.DocSearch-Commands-Key > svg[aria-label="Touche d\'échappement"]'),
       ).toBeInTheDocument();
       expect(
-        document.querySelector(
-          '.DocSearch-Commands-Key > svg[aria-label="Flèche vers le haut"]'
-        )
+        document.querySelector('.DocSearch-Commands-Key > svg[aria-label="Flèche vers le haut"]'),
       ).toBeInTheDocument();
+      expect(document.querySelector('.DocSearch-Commands-Key > svg[aria-label="Flèche le bas"]')).toBeInTheDocument();
       expect(
-        document.querySelector(
-          '.DocSearch-Commands-Key > svg[aria-label="Flèche le bas"]'
-        )
-      ).toBeInTheDocument();
-      expect(
-        document.querySelector(
-          '.DocSearch-Commands-Key > svg[aria-label="Touche d\'entrée"]'
-        )
+        document.querySelector('.DocSearch-Commands-Key > svg[aria-label="Touche d\'entrée"]'),
       ).toBeInTheDocument();
     });
   });
@@ -237,7 +213,7 @@ describe('api', () => {
               search: noResultSearch,
             };
           }}
-        />
+        />,
       );
 
       expect(document.querySelector(docSearchSelector)).toBeInTheDocument();
@@ -253,9 +229,7 @@ describe('api', () => {
       });
 
       expect(screen.getByText(/No results for/)).toBeInTheDocument();
-      expect(
-        document.querySelector('.DocSearch-Help a')
-      ).not.toBeInTheDocument();
+      expect(document.querySelector('.DocSearch-Help a')).not.toBeInTheDocument();
     });
 
     it('render the link to the repository', async () => {
@@ -267,10 +241,8 @@ describe('api', () => {
               search: noResultSearch,
             };
           }}
-          getMissingResultsUrl={({ query }) =>
-            `https://github.com/algolia/docsearch/issues/new?title=${query}`
-          }
-        />
+          getMissingResultsUrl={({ query }) => `https://github.com/algolia/docsearch/issues/new?title=${query}`}
+        />,
       );
 
       await act(async () => {
@@ -286,9 +258,7 @@ describe('api', () => {
       expect(screen.getByText(/No results for/)).toBeInTheDocument();
       const link = document.querySelector('.DocSearch-Help a');
       expect(link).toBeInTheDocument();
-      expect(link?.getAttribute('href')).toBe(
-        'https://github.com/algolia/docsearch/issues/new?title=q'
-      );
+      expect(link?.getAttribute('href')).toBe('https://github.com/algolia/docsearch/issues/new?title=q');
     });
   });
 });
