@@ -1,15 +1,9 @@
 import React from 'react';
 
-export function RecentIcon() {
+export function RecentIcon(): JSX.Element {
   return (
     <svg width="20" height="20" viewBox="0 0 20 20">
-      <g
-        stroke="currentColor"
-        fill="none"
-        fillRule="evenodd"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      >
+      <g stroke="currentColor" fill="none" fillRule="evenodd" strokeLinecap="round" strokeLinejoin="round">
         <path d="M3.18 6.6a8.23 8.23 0 1112.93 9.94h0a8.23 8.23 0 01-11.63 0" />
         <path d="M6.44 7.25H2.55V3.36M10.45 6v5.6M10.45 11.6L13 13" />
       </g>
