@@ -1,6 +1,6 @@
 import React from 'react';
 
-export function ResetIcon() {
+export function ResetIcon(): JSX.Element {
   return (
     <svg width="20" height="20" viewBox="0 0 20 20">
       <path
