@@ -1,14 +1,8 @@
 import React from 'react';
 
-export function SearchIcon() {
+export function SearchIcon(): JSX.Element {
   return (
-    <svg
-      width="20"
-      height="20"
-      className="DocSearch-Search-Icon"
-      viewBox="0 0 20 20"
-      aria-hidden="true"
-    >
+    <svg width="20" height="20" className="DocSearch-Search-Icon" viewBox="0 0 20 20" aria-hidden="true">
       <path
         d="M14.386 14.386l4.0877 4.0877-4.0877-4.0877c-2.9418 2.9419-7.7115 2.9419-10.6533 0-2.9419-2.9418-2.9419-7.7115 0-10.6533 2.9418-2.9419 7.7115-2.9419 10.6533 0 2.9419 2.9418 2.9419 7.7115 0 10.6533z"
         stroke="currentColor"
