@@ -1,6 +1,3 @@
 import type { DocSearchHit } from './DocSearchHit';
 
-export type StoredDocSearchHit = Omit<
-  DocSearchHit,
-  '_highlightResult' | '_snippetResult'
->;
+export type StoredDocSearchHit = Omit<DocSearchHit, '_highlightResult' | '_snippetResult'>;
