@@ -7,7 +7,7 @@ import { version } from './version';
 export function useSearchClient(
   appId: string,
   apiKey: string,
-  transformSearchClient: (searchClient: LiteClient) => LiteClient
+  transformSearchClient: (searchClient: LiteClient) => LiteClient,
 ): LiteClient {
   const searchClient = React.useMemo(() => {
     const client = liteClient(appId, apiKey);
@@ -18,10 +18,7 @@ export function useSearchClient(
     // it would also be sent on a DocSearch.js integration.
     // We therefore only add the `docsearch-react` user agent if `docsearch.js`
     // is not present.
-    if (
-      /docsearch.js \(.*\)/.test(client.transporter.algoliaAgent.value) ===
-      false
-    ) {
+    if (/docsearch.js \(.*\)/.test(client.transporter.algoliaAgent.value) === false) {
       client.addAlgoliaAgent('docsearch-react', version);
     }
 
