@@ -3,9 +3,7 @@ import type { DocSearchHit, InternalDocSearchHit } from '../types';
 const regexHighlightTags = /(<mark>|<\/mark>)/g;
 const regexHasHighlightTags = RegExp(regexHighlightTags.source);
 
-export function removeHighlightTags(
-  hit: DocSearchHit | InternalDocSearchHit
-): string {
+export function removeHighlightTags(hit: DocSearchHit | InternalDocSearchHit): string {
   const internalDocSearchHit = hit as InternalDocSearchHit;
 
   if (!internalDocSearchHit.__docsearch_parent && !hit._highlightResult) {
@@ -20,7 +18,5 @@ export function removeHighlightTags(
     return hit.hierarchy.lvl0;
   }
 
-  return lvl0.value && regexHasHighlightTags.test(lvl0.value)
-    ? lvl0.value.replace(regexHighlightTags, '')
-    : lvl0.value;
+  return lvl0.value && regexHasHighlightTags.test(lvl0.value) ? lvl0.value.replace(regexHighlightTags, '') : lvl0.value;
 }
